refactor: consolidate mock setups into a single file and remove unused mocks
