+++ frontend/src/pages/Schedule/Schedule/__tests__/table/modals/scheduleModal/addMode.test.tsx
@@ -27,16 +27,20 @@ import { MemoryRouter } from 'react-router-dom';
 import { vi, beforeEach, describe, test, expect } from 'vitest';
 
 import { mockUser } from '../../../fixtures/user';
-import { setupCampaignMocks } from '../../../mocks/campaignApi';
-import { setupCommandMocks } from '../../../mocks/commandApi';
-import { setupDatasetMocks } from '../../../mocks/datasetApi';
-import { ALWAYS_ONLY, mockDaypartRows, setupDaypartMocks } from '../../../mocks/daypartApi';
-import { setupEventApiMocks } from '../../../mocks/eventApi';
-import { setupLayoutsMocks } from '../../../mocks/layoutsApi';
-import { setupMediaMocks } from '../../../mocks/mediaApi';
-import { setupPlaylistMocks } from '../../../mocks/playlistApi';
-import { setupResolutionMocks } from '../../../mocks/resolutionApi';
-import { setupSyncGroupMocks } from '../../../mocks/syncGroupApi';
+import {
+  ALWAYS_ONLY,
+  mockDaypartRows,
+  setupCampaignMocks,
+  setupCommandMocks,
+  setupDatasetMocks,
+  setupDaypartMocks,
+  setupEventMocks,
+  setupLayoutsMocks,
+  setupMediaMocks,
+  setupPlaylistMocks,
+  setupResolutionMocks,
+  setupSyncGroupMocks,
+} from '../../../mocks/api';
 
 import { renderScheduleModal } from './helpers/renderScheduleModal';
 
@@ -118,7 +122,7 @@ describe('ScheduleEventModal - add mode (save & cancel)', () => {
     testQueryClient.clear();
     vi.clearAllMocks();
 
-    setupEventApiMocks();
+    setupEventMocks();
     setupDaypartMocks();
     setupResolutionMocks();
     setupLayoutsMocks();
